requirements.txt
pyserial
crccheck
paho-mqtt